ndt7: refactor code to easily plugin in the upload
This commit completes the yak shaving required to restructure
the code such that adding support for upload is easy.

We create a new ndt7/handler package where we do most of the
common work required by both download and upload.

This common code has been moved out of downlad, which now
looks very simple and understandable (IMHO).
